+++ projects/js-packages/partner-coupon/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.24 - 2022-09-08
+### Changed
+- Updated package dependencies.
+
+## 0.2.23 - 2022-08-30
+### Changed
+- Updated package dependencies. [#25814]
+
 ## 0.2.22 - 2022-08-23
 ### Changed
 - Updated package dependencies. [#25338, #25339, #25377, #25762]